Fix issue with mobile burger menu not showing up
Turns out I was using esbuild with a node version of 16 which made it incompatible and led to the below error

error esbuild@0.24.0: The engine "node" is incompatible with this module. Expected version ">=18". Got "16.13.1"

Upgrading to latest node version and setting that as a default and running the below steps helped fix things

- nvm install 22.9.0
- nvm alias default v22.9.0
- npm install --global yarn
- yarn
- rake assets:precompile